Add error rate metrics to LLMA daily aggregation
Changes:
- Change metric_value from UInt64 to Float64 to support percentages
- Add include_error_rates config flag (default: true)
- Add error rate calculation in SQL template
  - Checks for $ai_error property (non-empty, not 'null')
  - Checks for $ai_is_error = true
  - Calculates percentage rounded to 2 decimal places
- Update test scripts to handle Float64 values and error properties
- Add test events with configurable error rates

Error rates are stored as percentages (0-100) in the metrics table.
Metric names: {event_type}_error_rate (e.g., ai_trace_error_rate)

+++ posthog/clickhouse/migrations/0175_llma_metrics_daily.py
@@ -10,7 +10,7 @@ CREATE TABLE IF NOT EXISTS {LLMA_METRICS_DAILY_TABLE}
     date Date,
     team_id UInt64,
     metric_name String,
-    metric_value UInt64
+    metric_value Float64
 ) ENGINE = {MergeTreeEngine(LLMA_METRICS_DAILY_TABLE, replication_scheme=ReplicationScheme.REPLICATED)}
 PARTITION BY toYYYYMM(date)
 ORDER BY (team_id, date, metric_name)
